+++ android/app/build.gradle
@@ -2,6 +2,16 @@ apply plugin: "com.android.application"
 apply plugin: "org.jetbrains.kotlin.android"
 apply plugin: "com.facebook.react"
 
+
+project.ext.vectoricons = [
+    iconFontNames: [
+        'Feather.ttf',
+        'Ionicons.ttf',
+        'MaterialIcons.ttf',
+        'FontAwesome.ttf'
+    ]
+]
+
 /**
  * This is the configuration block to customize your React Native Android app.
  * By default you don't need to apply any configuration, just uncomment the lines you need.
@@ -130,3 +140,4 @@ configurations.all {
     }
 }
 
+apply from: "../../node_modules/react-native-vector-icons/fonts.gradle"
