+++ android/build.gradle
@@ -5,7 +5,7 @@ buildscript {
         compileSdkVersion = 36
         targetSdkVersion = 36
         ndkVersion = "27.1.12297006"
-        kotlinVersion = "2.1.20"
+        kotlinVersion = "1.8.22"
     }
     repositories {
         google()
@@ -18,11 +18,13 @@ buildscript {
     }
 }
 
+
 allprojects {
     repositories {
         google()
         mavenCentral()
     }
 }
+
 
 apply plugin: "com.facebook.react.rootproject"
