+++ android/settings.gradle
@@ -4,3 +4,4 @@ extensions.configure(com.facebook.react.ReactSettingsExtension){ ex -> ex.autoli
 rootProject.name = 'afamilia'
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
+
