+++ package.json
@@ -10,13 +10,17 @@
     "test": "jest"
   },
   "dependencies": {
-    "@react-native-community/blur": "^4.4.1",
     "@react-native/new-app-screen": "0.82.1",
+    "@react-navigation/native": "^7.1.21",
+    "@react-navigation/native-stack": "^7.8.0",
     "react": "19.1.1",
     "react-native": "0.82.1",
     "react-native-linear-gradient": "^2.8.3",
-    "react-native-safe-area-context": "^5.5.2",
-    "react-native-splash-screen": "^3.3.0"
+    "react-native-safe-area-context": "^5.6.2",
+    "react-native-screens": "^4.18.0",
+    "react-native-splash-screen": "^3.3.0",
+    "react-native-swiper": "^1.6.0",
+    "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
@@ -30,7 +34,8 @@
     "@react-native/metro-config": "0.82.1",
     "@react-native/typescript-config": "0.82.1",
     "@types/jest": "^29.5.13",
-    "@types/react": "^19.1.1",
+    "@types/react": "^19.2.7",
+    "@types/react-native": "^0.73.0",
     "@types/react-test-renderer": "^19.1.0",
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
